requirements.txt
# Core dependencies
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
scipy==1.11.4

# API Framework
flask==3.0.0
flask-cors==4.0.0

# Document Processing
PyPDF2==3.0.1
openpyxl==3.1.2

# Data visualization
matplotlib==3.8.2
seaborn==0.13.0

# Machine Learning
joblib==1.3.2
xgboost==2.0.3

# Google Cloud (for deployment)
google-cloud-bigquery==3.14.1
google-cloud-storage==2.14.0
google-cloud-pubsub==2.19.0
google-cloud-logging==3.9.0
google-cloud-secret-manager==2.17.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
